+++ releaf-core/spec/lib/releaf/core/application_spec.rb
@@ -1,0 +1,10 @@
+require "spec_helper"
+
+describe Releaf::Core::Application do
+  describe "#configure" do
+  end
+
+  describe "#render_layout" do
+  end
+end
+