+++ releaf-i18n_database/app/builders/releaf/i18n_database/translations/table_builder.rb
@@ -1,11 +1,11 @@
 module Releaf::I18nDatabase::Translations
   class TableBuilder < Releaf::Builders::TableBuilder
     def column_names
-      [:key] + Releaf.all_locales
+      [:key] + Releaf.application.config.all_locales
     end
 
     def head_cell_content(column)
-      if Releaf.all_locales.include? column.to_s
+      if Releaf.application.config.all_locales.include? column.to_s
         translate_locale(column)
       else
         super
@@ -21,7 +21,7 @@ module Releaf::I18nDatabase::Translations
     end
 
     def cell_format_method(column)
-      if Releaf.all_locales.include? column
+      if Releaf.application.config.all_locales.include? column
         :locale_value
       else
         super